finished - total 4-6 hours of work
Everything seems to be working. I dont want to take the extra time to
see if i can save just the object itself to the file and decided to keep
the save the way serj has tought me. no errors listed and runs from
start to finish new order or load order